requirements.txt
requests>=0.10,<3.0.0
six>=1.0.0,<2.0.0
wheel==0.30.0
boto==2.48.0
python-dateutil==2.6.1
diskcache==2.9.0
setuptools>=39.0.1
dataclasses==0.5
